requirements.txt
click==7.1.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.61.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
passlib==1.7.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.20.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.35
starlette==0.13.6
typing_extensions==4.1.1
uvicorn==0.12.2
